refactor: create component phone input mask
